[Feat] - 결과/스킬 페이지 컴포넌트 분리
